refactor: login form
-Move login form inputs component into Login Form since it isn't being reused. -Delete login form inputs component.
