+++ client/src/components/forms/shared/LoginForm.tsx
@@ -1,10 +1,9 @@
+//LoginForm.tsx
+
 import { Link } from "react-router-dom";
 import type { AuthFormData } from "../../../pages/Login";
-import { useLayoutEffect, useState } from "react";
-import loadable from "@loadable/component";
-import Turnstile from "react-turnstile";
-
-const LoginFormInputs = loadable(() => import("./LoginFormInputs"));
+import { useState } from "react";
+import styles from "./styles/LoginFormInputs.module.css";
 
 interface PropTypes {
   formData: AuthFormData;
@@ -14,6 +13,66 @@ interface PropTypes {
   setGuestLogin?: (value: boolean) => void;
   serverError: string;
 }
+
+type FormInputProps = {
+  inputData: { [key: string]: string | boolean | null };
+  inputValues: { [key: string]: string };
+  setInputValues: (value: { [key: string]: string }) => void;
+};
+
+declare module "react" {
+  interface HTMLAttributes<T> extends AriaAttributes, DOMAttributes<T> {
+    focused?: string; //Allows for custom HTML attribute type called focused
+  }
+}
+// Used by LoginForm.tsx component
+function LoginFormInputs({
+  inputData,
+  inputValues,
+  setInputValues,
+}: FormInputProps) {
+  const [focused, setFocused] = useState<boolean>(false);
+  const { pattern, asterisk: dispAsterisk, ...inputs } = inputData;
+
+  const handleChange = (e: React.FormEvent<HTMLInputElement>) => {
+    setInputValues({
+      ...inputValues,
+      [e.currentTarget.name]: e.currentTarget.value,
+    });
+  };
+
+  return (
+    <>
+      <label
+        htmlFor={inputData.id?.toString()}
+        className="relative mr-auto cursor-pointer pl-1 hover:border-0"
+      >
+        {dispAsterisk ? `${inputData.label} *` : inputData.label}
+      </label>
+      <input
+        {...inputs}
+        pattern={
+          inputData.name?.toString().startsWith("confirm")
+            ? inputValues.password
+            : inputData.name?.toString().startsWith("email")
+              ? undefined
+              : pattern?.toString()
+        }
+        className="relative rounded-md border-2 border-solid p-2 pl-4"
+        onChange={handleChange}
+        onBlur={() => setFocused(true)}
+        onFocus={() => setFocused(false)}
+        focused={focused.toString()}
+      />
+      <span
+        className={`${styles.error} relative hidden items-center justify-center text-center text-sm`}
+      >
+        {inputData.err}
+      </span>
+    </>
+  );
+}
+
 // Used by Login.tsx and Register.tsx components
 function LoginForm({
   formData,
@@ -23,15 +82,10 @@ function LoginForm({
   setGuestLogin,
   serverError,
 }: PropTypes) {
-  const [captchaToken, setCaptchaToken] = useState("");
-
-  useLayoutEffect(() => {
-    LoginFormInputs.load();
-  }, []);
 
   return (
     <form
-      onSubmit={captchaToken ? submitForm : () => {} }
+      onSubmit={submitForm}
       className="relative mx-5 flex w-full  max-w-md flex-col gap-4 font-nunito text-xl"
     >
       {formData.map((data) => (
@@ -67,17 +121,6 @@ function LoginForm({
           </div>
         </div>
       )}
-      <div
-        aria-label="Cloudflare Turnstile Captcha Verification"
-        className="flex w-full items-center justify-center pt-2"
-      >
-        <Turnstile
-          sitekey="0x4AAAAAAAcX0OWvMBA9t7JC"
-          onVerify={(token) => {
-            setCaptchaToken(token);
-          }}
-        />
-      </div>
       <button
         type="submit"
         className="text-md mt-3 flex w-full items-center justify-center rounded-lg border-2 bg-sky-700 py-4 text-white outline-green-900 hover:scale-[1.01] hover:brightness-105"
